api/requirements.txt
anyio==3.7.1
bcrypt==3.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
email-validator==1.2.1
exceptiongroup==1.0.4
fastapi==0.104.1
greenlet==3.0.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpcore==0.15.0
hyperframe==6.0.1
idna==3.4
packaging==23.1
pip==23.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.13
PyJWT==2.4.0
pyserial==3.5
python-dotenv==0.21.0
python-multipart==0.0.6
redis==5.0.1
regex==2023.10.3
setuptools==68.0.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.21
sse-starlette==2.1.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.20.0
wheel==0.41.2
